fix: extract only sequence number from invoice for max calculation
Changed SUBSTRING position from 4 to 12 to extract only the 4-digit
sequence number (e.g., "0001") instead of including the date prefix.
This prevents the date from being duplicated in new invoice numbers.

Invoice format: INV + YYYYMMDD + XXXX (e.g., INV202601090001)
- Position 1-3: "INV"
- Position 4-11: "20260109" (date)
- Position 12-15: "0001" (sequence)

+++ shina-magazin-api/src/main/java/uz/shinamagazin/api/service/SaleService.java
@@ -238,8 +238,8 @@ public class SaleService {
 
     private String generateInvoiceNumber() {
         String prefix = "INV" + LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
-        Long maxNumber = saleRepository.findMaxInvoiceNumber(prefix);
-        long nextNumber = (maxNumber != null ? maxNumber : 0L) + 1;
+        Integer maxNumber = saleRepository.findMaxInvoiceNumber(prefix);
+        int nextNumber = (maxNumber != null ? maxNumber : 0) + 1;
         return prefix + String.format("%04d", nextNumber);
     }
